Enable opening RedNotebook on the specified date via the command line option -s
The date format must be supplied in the format YYYY-MM-DD. e.g. with the long option type:
 rednotebook --start-date=2013-09-25

This is because of using the existing string to date conversion routine.
Adding use of time.strptime() would be onerous as then the user would need to supply the time format as well.
Thus I have chosen to follow a KISS principal.

Amend the man page with the new option.