Implement new search: support multiple words
Previously, a search would only return results with exact matches, that is if
the search string is a substring of the text in the window buffer.  With this
new edition, we support searching for multiple words. A result would only be
shown if *all* of those words appear at least once in the text.