rednotebook_merge: Avoid duplicate data when merging
It is useful to be able to merge notebooks regularly across multiple
machines. The existing script adds the new material again, even if a
previous merge already added it.

Add a check for this, so that repeated merging will not create lots of
duplicate text.

Also include a summary of what happened and mention a script dependency.

Signed-off-by: Simon Glass <sjg@chromium.org>