Declare Python dependencies in setup.py (#871)
* Declare dependencies in setup.py

This makes it easier to install the application with a simple `pip install .`, or `pipx install .` for a global installation in a separate venv.

pyenchant was added as an optional dependency in the "spellcheck" extra group.

* Update instructions for macOS installation

Simplify installation instructions for macOS using pipx, which
will automatically set up a virtual environment and create a link
to the application on the system PATH.

This way there is no risk of screwing up other Python packages
installed by the user.

* CI/CD: let the venv use system packages

This way we test against Python dependencies packaged
with Ubuntu.